Test cases addition and minor refactoring.
