+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,6 +1,5 @@
 package ImageHoster.controller;
 
-import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
@@ -10,12 +9,14 @@ import ImageHoster.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.time.LocalDate;
 import java.util.*;
 
 @Controller
@@ -26,9 +27,6 @@ public class ImageController {
 
     @Autowired
     private TagService tagService;
-
-    @Autowired
-    private CommentService commentService;
 
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
@@ -87,20 +85,6 @@ public class ImageController {
         newImage.setDate(new Date());
         imageService.uploadImage(newImage);
         return "redirect:/images";
-    }
-
-
-    @RequestMapping(value = "/image/{imageId}/{imageTitle}/comments", method = RequestMethod.POST)
-    public String uploadAndGetComments(@PathVariable("imageId") Integer imageId, @PathVariable("imageTitle") String imageTitle, @RequestParam("comment") String comments, HttpSession session) {
-        Image image = imageService.getImage(imageId);
-        User user = (User) session.getAttribute("loggeduser");
-        Comment comment = new Comment();
-        comment.setText(comments);
-        comment.setDate(LocalDate.now());
-        comment.setUser(user);
-        comment.setImage(image);
-        commentService.uploadComments(comment);
-        return "redirect:/images/" + imageId + "/" + imageTitle;
     }
 
     //This controller method is called when the request pattern is of type 'editImage'
