+++ src/test/java/ImageHoster/controller/HomeControllerTest.java
@@ -1,4 +1,3 @@
-/*
 package ImageHoster.controller;
 
 import ImageHoster.service.ImageService;
@@ -36,4 +35,3 @@ public class HomeControllerTest {
                 .andExpect(content().string(containsString("Image Hoster")));
     }
 }
-*/
